+++ python/lsst/__init__.py
@@ -1,0 +1,4 @@
+import pkgutil
+
+__path__ = pkgutil.extend_path(__path__, __name__)
+
