+++ python/lsst/multiprofit/asinhstretchsigned.py
