+++ python/lsst/multiprofit/componentconfig.py
