+++ python/lsst/multiprofit/config.py
