+++ python/lsst/multiprofit/limits.py
