+++ python/lsst/multiprofit/modeller.py
