+++ python/lsst/multiprofit/plots.py
