+++ python/lsst/multiprofit/priors.py
