+++ python/lsst/multiprofit/transforms.py
