+++ python/lsst/multiprofit/utils.py
